Fix buffer leaks in MP codec and in tests (#15)
Motivation:

So far, the LeakDetector was not used when running the tests (with the -P leak profile), because -Dio.netty5.buffer.leakDetectionEnabled=true was missing in the test.argLine property of the leak profile (in the parent pom).

Now, using the option, many leaks are reported. This PR fixes the remaining leaks and will unblock the other #14 PR which is currently failing because of the buffer leaks.

These leaks only concerns this MP codec, not the old MP from the old Netty 4.x, because in old version, the MP was using some unpooled buffers, and in 4.x, the not-released unpooled buffers are not logged by the LeakDetector.
Now, we are only used pooled buffers, which should all be properly closed.

Motifications:

In the source codec:

fixed AbstractMemoryHttpData.setContentInternal which now verifies if the current resource content must be closed. Before, the MemoryAttribute.prepareSend method was calling setContentInternal and then previous content resource was not closed
fixed HttpPostMultipartRequestDecoder class: The attributes from the currentFieldAttributes list are now closed before emptying the currentFieldAttributes list.
fixed HttpPostRequestEncoder class: empty chunk buffers are now closed
fixed HttpPostStandardRequestDecoder: setFinalBuffer is now closing the current attribute buffer (if the attribute is not a memory based (because for memory http data, the getBuffer() method returns a singleton, and does not allocate a copy of the current memory http data buffer)
in the test classes:

fixed many tests where buffers and codec objects were not closed properly.

Result:

The CI should now pass when the -P leak profile is used, and the other #14 PR will be unblocked by this one (the other #14 PR will have to be rebased on top of main, once this PR will be merged).