Simplify buffer ownership for HttpData getBuffer/getContent methods (#17)
Motivation:

Currently, the buffer ownership of the buffers returned by the HttpData getBuffer/content methods depends on the nature of the HttpData:

for disk based http data, a copy is returned by the getBuffer()/content() methods; so the user is getting the buffer ownership and is assumed to close the returned buffer.
but for memory based http data , the internal buffer is returned by the getBuffer()/content() methods; so the user don't have to close the returned buffer, because it will be closed when the http data object will be closed. However, the user can still close the buffer, in this case it won't be closed a second time when the http data object is closed.

Modifications:

This PR eliminate the difference between disk based vs memory based attribute, by replacing the HttpData.getBuffer() method with a "usingBuffer<ThrowingConsumer<Buffer>>" callback. So the ownership is left to the HttpData and is not transferred to the caller, and when the callback returns, then the passed buffer will be closed in case the HttpData is disk based.
likewise, the HttpData.content() is replaced by a "usingContent<ThrowingConsumer<Buffer>>" callback.

The callback is a "Helpers.ThrowingConsumer", which is similar to java Consumer except that the consumer may (or may not) throw exceptions.

Results:

the contract of the usingBuffer/usingContent methods is clearer, and the user knows that it does not acquire ownership for the buffer passed as parameter. And if the user wants to obtain the ownership, he can then call the split/copy or send method of the buffer passed to the withBuffer/withContent callbacks.