Decode percent encoding on-the-fly to avoid HttpData.usingBuffer Motivation:
HttpData.usingBuffer may require copying data when the data is only present on disk. This is slow and defeats the purpose of moving the data to disk in the first place. Likely most large files will use multipart where this is not an issue, but the possibility still makes alternative HttpData implementations awkward. Thus, usingBuffer should be avoided where possible.

Modification:

Alter HttpPostStandardRequestDecoder to decode data on-the-fly instead of all at once.

Result:

usingBuffer is not called anymore. usingBuffer is now unused outside the HttpData implementations themselves.